refactor: API 관련 상수 리폴더링
